refactor(apps/web): info content subdividion in components for better code maintenance
