+++ apps/web/src/pages/private/Layout/components/Info/components/Credits/components/CreditArtist/index.tsx
@@ -1,0 +1,24 @@
+import { useTheme } from "styled-components";
+import { Chip } from "../../../../../../../../../components/Chip";
+import { Typography } from "../../../../../../../../../components/Typography";
+import { CreditArtistContainer, CreditArtistInfo } from "./styles";
+
+function CreditArtist() {
+  const theme = useTheme();
+
+  return (
+    <CreditArtistContainer>
+      <CreditArtistInfo>
+        <Typography variant="subtitle" size={14}>
+          Artist name
+        </Typography>
+        <Typography variant="text" size={12} color={theme.text.subdued}>
+          Role
+        </Typography>
+      </CreditArtistInfo>
+      <Chip />
+    </CreditArtistContainer>
+  );
+}
+
+export { CreditArtist };
