+++ apps/web/src/pages/private/Layout/components/Info/components/Credits/components/CreditArtist/styles.ts
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+const CreditArtistContainer = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 18px;
+`;
+
+const CreditArtistInfo = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+  justify-content: space-between;
+`;
+
+export { CreditArtistContainer, CreditArtistInfo };
