+++ apps/web/src/pages/private/Layout/components/Info/components/Credits/index.tsx
@@ -1,0 +1,26 @@
+import { useTheme } from "styled-components";
+import { Typography } from "../../../../../../../components/Typography";
+import { SectionPanel, SectionPanelHeader } from "../../styles";
+import { CreditArtist } from "./components/CreditArtist";
+
+function Credits() {
+  const theme = useTheme();
+
+  return (
+    <SectionPanel>
+      <SectionPanelHeader>
+        <Typography variant="title" size={14}>
+          Credits
+        </Typography>
+        <Typography variant="link" size={14} color={theme.text.subdued}>
+          Show all
+        </Typography>
+      </SectionPanelHeader>
+      <CreditArtist />
+      <CreditArtist />
+      <CreditArtist />
+    </SectionPanel>
+  );
+}
+
+export { Credits };
