+++ apps/web/src/pages/private/Layout/components/Info/components/Header/styles.ts
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+
+const HeaderContainer = styled.header`
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 30px 12px 16px;
+`;
+
+const HeaderButtons = styled.div`
+  display: flex;
+  gap: 12px;
+`;
+
+export { HeaderButtons, HeaderContainer };
