+++ apps/web/src/pages/private/Layout/components/Info/components/Tour/index.tsx
@@ -1,0 +1,25 @@
+import { useTheme } from "styled-components";
+import { Typography } from "../../../../../../../components/Typography";
+import { SectionPanel, SectionPanelHeader } from "../../styles";
+import { TourDate } from "./components/TourDate";
+
+function Tour() {
+  const theme = useTheme();
+
+  return (
+    <SectionPanel>
+      <SectionPanelHeader>
+        <Typography variant="title" size={14}>
+          On tour
+        </Typography>
+        <Typography variant="link" size={14} color={theme.text.subdued}>
+          Show all
+        </Typography>
+      </SectionPanelHeader>
+      <TourDate />
+      <TourDate />
+    </SectionPanel>
+  );
+}
+
+export { Tour };
